requirements.txt
aiohttp==3.11.11
asyncpg==0.30.0
apache-airflow==2.10.4
python-dotenv==1.0.1
python-dateutil==2.9.0.post0